Sample Data and Matlab Code
This study uses Langland's formula to combine all the factors, including the mass, charge, isoelectric point, hydrophilic and hydrophobic properties, and equipotential of the amino acids to analyze the flu proteins. The hydrophilic and hydrophobic
properties express the ability of viral proteins to interact with sialic acid. There is a linear relationship between the isoelectric point of the amino acids and the pH value of the amino acids. The isoelectric point and pH value of the amino acid sequences change the equivalent charge potential. There is also a linear relationship between hydrophilic and hydrophobic properties of the amino acids and the heat or intrinsic energy level. The hydrophobicity and the heat/ intrinsic energy change the equivalent mass. Thus, we use the equivalent extended mass and charge value as the ratio to calculate the overall Mass-Charge ratio. To compare the virus with different length, we further use the sub semi variance of the entire sequence. To amplify the high momentum, we use 11th fractional variance to generate the loglog complex spectrum, to better understand the binding dynamics. Eventually, we want to investigate why duck can genetically survive the most bird flus.